+++ pkg/controller/add_linstorcontroller.go
@@ -18,10 +18,10 @@ limitations under the License.
 package controller
 
 import (
-	"github.com/piraeusdatastore/piraeus-operator/pkg/controller/linstornodeset"
+	"github.com/piraeusdatastore/piraeus-operator/pkg/controller/linstorcontroller"
 )
 
 func init() {
 	// AddToManagerFuncs is a list of functions to create controllers and add them to a manager.
-	AddToManagerFuncs = append(AddToManagerFuncs, linstornodeset.Add)
+	AddToManagerFuncs = append(AddToManagerFuncs, linstorcontroller.Add)
 }