+++ pkg/controller/linstorcontroller/linstorcontroller_controller_test.go
@@ -1,4 +1,4 @@
-package linstorcontrollerset
+package linstorcontroller
 
 import (
 	"reflect"
@@ -12,17 +12,17 @@ import (
 func TestNewConfigMapForPCS(t *testing.T) {
 	testcases := []struct {
 		name     string
-		spec     *piraeusv1alpha1.LinstorControllerSet
+		spec     *piraeusv1alpha1.LinstorController
 		expected *corev1.ConfigMap
 	}{
 		{
 			name: "default-settings",
-			spec: &piraeusv1alpha1.LinstorControllerSet{
+			spec: &piraeusv1alpha1.LinstorController{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default-ns",
 				},
-				Spec: piraeusv1alpha1.LinstorControllerSetSpec{
+				Spec: piraeusv1alpha1.LinstorControllerSpec{
 					DBConnectionURL:     "etcd://etcd.svc:5000/",
 					DBCertSecret:        "",
 					LinstorClientConfig: piraeusv1alpha1.LinstorClientConfig{},
@@ -59,12 +59,12 @@ controllers = http://test.default-ns.svc:3370
 		},
 		{
 			name: "with-ssl-without-client-cert",
-			spec: &piraeusv1alpha1.LinstorControllerSet{
+			spec: &piraeusv1alpha1.LinstorController{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default-ns",
 				},
-				Spec: piraeusv1alpha1.LinstorControllerSetSpec{
+				Spec: piraeusv1alpha1.LinstorControllerSpec{
 					DBConnectionURL:     "etcd://secure.etcd.svc:443/",
 					DBCertSecret:        "mysecret",
 					LinstorClientConfig: piraeusv1alpha1.LinstorClientConfig{},
@@ -102,12 +102,12 @@ controllers = http://test.default-ns.svc:3370
 		},
 		{
 			name: "with-ssl-with-client-cert",
-			spec: &piraeusv1alpha1.LinstorControllerSet{
+			spec: &piraeusv1alpha1.LinstorController{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default-ns",
 				},
-				Spec: piraeusv1alpha1.LinstorControllerSetSpec{
+				Spec: piraeusv1alpha1.LinstorControllerSpec{
 					DBConnectionURL:     "etcd://secure.etcd.svc:443/",
 					DBCertSecret:        "mysecret",
 					DBUseClientCert:     true,
@@ -148,12 +148,12 @@ controllers = http://test.default-ns.svc:3370
 		},
 		{
 			name: "with-https-auth",
-			spec: &piraeusv1alpha1.LinstorControllerSet{
+			spec: &piraeusv1alpha1.LinstorController{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "test",
 					Namespace: "default-ns",
 				},
-				Spec: piraeusv1alpha1.LinstorControllerSetSpec{
+				Spec: piraeusv1alpha1.LinstorControllerSpec{
 					DBConnectionURL:              "etcd://etcd.svc:5000/",
 					DBCertSecret:                 "",
 					LinstorHttpsControllerSecret: "controller-secret",